fix(server_info): rippled-conformance pass on #480
Addresses review of PR #488:

- ledger age threshold is now 1,000,000s (rippled NetworkOPs.cpp:2951),
  not 600s — short stalls are no longer hidden behind age=0.
- server_state_duration_us now reports time-in-current-mode (rippled
  emits `current.count()` at NetworkOPs.cpp:4846), not process uptime.
- initial_sync_duration_us is captured on first transition into Full
  and surfaced when non-zero (NetworkOPs.cpp:4814-4820, 4847-4848).
- Human-mode load_factor_fee_escalation / _queue emitted as
  decimal-from-reference ratios when escalation is active, matching
  NetworkOPs.cpp:2902-2912.
- closed_ledger.age omitted when the close time is in the future,
  mirroring rippled NetworkOPs.cpp:2962-2969.
- stateAccounting holds the adaptor mutex across transition() so the
  operatingMode field and accounting tracker observe the same
  serialization order.
- state-accounting keys ordered by OperatingMode index to mirror
  rippled's iteration in NetworkOPs.cpp:871-872 + 4837-4845.
- Comments on jq_trans_overflow / peer_disconnects_resources / the
  ping-timeout stand-in honestly call out the semantic gap with
  rippled's JobQueue / ResourceManager counters.
- Drop the unnecessary uint32() cast on TxCount() (lint unconvert).

Adds focused tests for the high-age threshold, the closed_ledger
skew-omission, the snapshot-driven top-level durations, and the
initial-sync capture on first Full transition.

+++ internal/ledger/service/service.go
@@ -1361,7 +1361,7 @@ func (s *Service) GetTxQMetrics() txq.Metrics {
 	}
 	var txInLedger uint32
 	if s.openLedger != nil {
-		txInLedger = uint32(s.openLedger.TxCount())
+		txInLedger = s.openLedger.TxCount()
 	}
 	return s.txQueue.GetMetrics(txInLedger)
 }
